Offline finished and tested with given input

+++ SymbolTable.cpp
@@ -3,6 +3,7 @@
 #include<iostream>
 #include <cstdlib>
 #include <vector>
+#include <fstream>
 
 using namespace std;
 
@@ -38,7 +39,7 @@ private:
 
         for(int i=0;i<name.length(); i++){
             pos  = (pos *31 + name[i]) % n;
-            cout <<name << " pos " << pos << endl;
+            //cout <<name << " pos " << pos << endl;
         }
 
         return (int)pos;
@@ -48,14 +49,14 @@ public:
     //variables
     ScopeTable * parentScope;
 
-    ScopeTable(int x = 10 , int id){
+    ScopeTable(int id , int x = 10){
         n = x;
         this->id = id;
         scopeTable = new SymbolInfo*[n];
         for(int i=0; i<n; i++){
             scopeTable[i] = NULL;
         }
-
+        cout << "ScopeTable with id "<< id << " is created\n";
     }
 
     void setId(int id){
@@ -72,22 +73,26 @@ public:
         newItem->setType(type);
         newItem->next = NULL;
 
+        int chainPos = 0;
         int pos = hashFunction(name);
-        SymbolInfo *cur = scopeTable[pos];
+        SymbolInfo *prev,*cur = scopeTable[pos];
         if(cur == NULL)
             scopeTable[pos] = newItem;
         else{
-
-            while(cur->next != NULL){
+            prev = NULL;
+            while(cur != NULL){
                 if(cur->getName() == name){
+                    cout <<"<" << name << "," <<type<<"> already exists\n";
                     return false;
                 }
+                chainPos++;
                 //prev = cur;
+                prev = cur;
                 cur = cur->next;
             }
-            cur->next = newItem;
+            prev->next = newItem;
         }
-        cout << "<" << name << "," << type << "> inserted at "<<pos << endl;
+        cout << "<" << name << "," << type << "> inserted at ScopeTable#"<<id << " at ("<<pos << ","<<chainPos << ")\n";
         return true;
 
     }
@@ -95,24 +100,31 @@ public:
     SymbolInfo * lookup(string name){
         SymbolInfo * cur = new SymbolInfo;
 
-
+        int chainPos = 0;
         int pos = hashFunction(name);
         cur = scopeTable[pos];
         while(cur!=NULL){
-            if(cur->getName() == name)
+            if(cur->getName() == name){
+                cout << "<" << name << "," << cur->getType() << "> found in ScopeTable#"<<id << " at ("<<pos << ","<<chainPos << ")\n";
                 return cur;
+            }
+
             cur = cur->next;
         }
+
+        cout << "Item not found\n";
+
         return cur;
 
     }
 
     bool deleteItem(string name){
         SymbolInfo *prev,* cur ;
-
+        lookup(name);
 
         int pos = hashFunction(name);
         cur = scopeTable[pos];
+        int chainPos = 0;
 
         prev = NULL;
         while(cur != NULL){
@@ -122,18 +134,19 @@ public:
                 else
                     scopeTable[pos] = cur->next;
                 delete cur;
+                cout << "Deleted entry of ScopeTable#" << id << " at position ("<<pos << "," << chainPos << ")\n";
                 return true;
             }
+            chainPos++;
             prev = cur;
             cur = cur->next;
-
         }
         return false;
     }
 
     void print(){
         SymbolInfo *cur;
-        cout << n << endl;
+        cout << "ScopeTable#"<<id << ":"<< endl;
         for(int i=0;i<n;i++){
             printf("Bucket%d :",i);
             cur = scopeTable[i];
@@ -156,6 +169,7 @@ public:
                 delete cur;
             }
         }
+        cout << "ScopeTable with id "<< id << " is deleted\n";
         delete []scopeTable;
     }
 
@@ -173,11 +187,12 @@ public:
         currentScope = NULL;
         curId = 0 ;
         bucketSize = n;
+        enterScope();
     }
 
     void enterScope(){
         curId++;
-        ScopeTable * newScope = new ScopeTable(bucketSize,curId);
+        ScopeTable * newScope = new ScopeTable(curId,bucketSize);
         newScope -> parentScope = currentScope;
         currentScope = newScope;
     }
@@ -191,12 +206,16 @@ public:
     }
 
     bool insert(string name, string type){
+        if(currentScope == NULL)
+            return false;
         if(currentScope->insert(name,type))
             return true;
         else return false;
     }
 
     bool remove(string name){
+        if(currentScope == NULL )
+            return false;
         if(currentScope->deleteItem(name))
             return true;
         else return false;
@@ -208,19 +227,23 @@ public:
         while(cur!=NULL){
             result = cur->lookup(name);
             if(result != NULL)
-                return res;
+                return result;
             else cur = cur->parentScope;
         }
-        return res;
+        return result;
 
     }
 
     void printCurrentScope(){
-        currentScope->print();
+        if(currentScope != NULL)
+            currentScope->print();
+        else
+            cout << "No scope table created\n";
     }
     void printAll(){
         ScopeTable * cur = currentScope;
-
+        if(cur == NULL)
+            cout << "No Scope Table!!\n";
         while (cur!=NULL){
             cur->print();
             cur = cur->parentScope;
@@ -239,42 +262,72 @@ public:
 };
 
 
+
 int main(){
-    freopen("in.txt","rw",stdin);
-    int n,x;
-    cin >> n;
-    ScopeTable table(n);
+    ifstream in("in.txt");
+    ofstream out("out.txt");
+    streambuf *inbuf = cin.rdbuf();
+    cin.rdbuf(in.rdbuf());
+    //cout.rdbuf(out.rdbuf());
+    int n,f = 1;
+    in >> n;
+    //cout << n;
+    SymbolTable st(n);
+    string cmd;
     while(true){
-        if(scanf("%d",&x) == EOF){
-            freopen("/dev/tty","r",stdin);
+
+        if(cin.eof()){
+            cin.rdbuf(inbuf);
+            cout << "cin changed" << endl;
+            //break;
         }
-        if(x == 1){
-            string name,t;
-            cin >> name >> t;
-            table.insert(name,t);
+        cin >> cmd;
+        cout << cmd << " ";
+        if(cmd == "S"){
+            cout << endl <<"\t";
+            st.enterScope();
         }
-        if(x == 2){
+        if(cmd == "E"){
+            cout << endl << "\t";
+            st.exitScope();
+        }
+        if(cmd == "I"){
+            string name,type;
+            cin >> name >> type;
+            cout << name << " "<<type << endl << "\t";
+            st.insert(name,type);
+        }
+        if(cmd == "D"){
             string name;
             cin >> name;
-            SymbolInfo * tmp = table.lookup(name);
-            if(tmp != NULL)
-                cout << tmp->getName() << "," <<tmp->getType() << endl;
-            else
-                cout << "Item was not found!!\n";
+            cout << name << endl << "\t";
+            st.remove(name);
         }
-        if(x == 3){
+        if(cmd == "L"){
             string name;
             cin >> name;
-            bool b = table.deleteItem(name);
-            if(b)
-                cout << "item deleted successfully\n";
-            else
-                cout << "Item was not found!!\n";
+            cout << name << endl << "\t";
+            st.lookup(name);
         }
-        if(x == 4)
-            table.print();
-        if(x == 5)
+        if(cmd == "P"){
+            string printType;
+            cin >> printType;
+            cout << printType << endl << "\t";
+            if(printType == "A"){
+                st.printAll();
+            }
+            else if(printType == "C"){
+                st.printCurrentScope();
+            }
+        }
+        if(cmd == "Q"){
+            cout << endl << "\t";
             break;
+        }
+
+        //break;
+
+
     }
     return 0;
 }
